rename: 구조 변경 추가 MainBox -> MainBox/Genres
