+++ src/screens/MainBox/Genres/Components/DaysTop.js
@@ -1,5 +1,5 @@
-import { createExtendsRelation } from "../../../utils.js";
-import Component from "../../Component.js";
+import { createExtendsRelation } from "../../../../utils.js";
+import Component from "../../../Component.js";
 
 function DaysTop(infoObject) {
   Component.call(this, infoObject);
