+++ src/screens/MainBox/Genres/Components/MainBanner.js
@@ -1,7 +1,7 @@
-import { createExtendsRelation } from "../../../utils.js";
-import Component from "../../Component.js";
+import { createExtendsRelation } from "../../../../utils.js";
+import Component from "../../../Component.js";
+import carousel from "../../../../modules/carousel.js";
 import BannerImage from "./BannerImage.js";
-import carousel from "../../../modules/carousel.js";
 
 function MainBanner(infoObject) {
   Component.call(this, infoObject);
