+++ src/screens/MainBox/Genres/Webtoon/genre/BoyGenre.js
@@ -1,5 +1,5 @@
-import Component from "../../../Component.js";
-import { createExtendsRelation } from "../../../../utils.js";
+import Component from "../../../../Component.js";
+import { createExtendsRelation } from "../../../../../utils.js";
 
 function BoyGenre(target) {
   Component.call(this, target);
