+++ src/screens/MainBox/Genres/components.js
@@ -1,0 +1,23 @@
+import BigCardList from "./Components/BigCardList.js";
+import DateTop from "./Components/DateTop.js";
+import DaysTop from "./Components/DaysTop.js";
+import FullButton from "./Components/FullButton.js";
+import GenreTop from "./Components/GenreTop.js";
+import MainBanner from "./Components/MainBanner.js";
+import NavDetail from "./Components/NavDetail";
+import RecommendEvent from "./Components/RecommendEvent.js";
+import SubBanner from "./Components/SubBanner.js";
+
+const components = {
+  mainBanner: MainBanner,
+  navDetail: NavDetail,
+  subBanner: SubBanner,
+  daysTop: DaysTop,
+  bigCardList: BigCardList,
+  genreTop: GenreTop,
+  dateTop: DateTop,
+  recommendEvent: RecommendEvent,
+  fullButton: FullButton,
+};
+
+export default components;
