--- src/screens/MainBox/components.js
@@ -1,23 +1,0 @@
-import BigCardList from "./Components/BigCardList";
-import DateTop from "./Components/DateTop";
-import DaysTop from "./Components/DaysTop";
-import FullButton from "./Components/FullButton";
-import GenreTop from "./Components/GenreTop";
-import MainBanner from "./Components/MainBanner";
-import NavDetail from "./Components/NavDetail";
-import RecommendEvent from "./Components/RecommendEvent";
-import SubBanner from "./Components/SubBanner";
-
-const components = {
-  mainBanner: MainBanner,
-  navDetail: NavDetail,
-  subBanner: SubBanner,
-  daysTop: DaysTop,
-  bigCardList: BigCardList,
-  genreTop: GenreTop,
-  dateTop: DateTop,
-  recommendEvent: RecommendEvent,
-  fullButton: FullButton,
-};
-
-export default components;
